First replacement of OpenSSL ASN1 code: X509_ALGOR and X509_EXTENSION structures.
This turned up loads of bugs in the encoder
and decoder which are now fixed. Mainly due
to mishandling of SEQUENCEs in the ASN1_ANY
type.

It also removes an issue with the "netscape_hack"
field of X509_EXTENSION. Since the critical
flag is DEFAULT FALSE it will be omitted if it
is FALSE. However some broken certificates encode
it anyway for FALSE so historically the "netscape_hack"
flag was set to mean "encode even if its FALSE". Now
however the ASN1_BOOLEAN type has three values > 0
for TRUE (of which 0xff is the correct DER value)
== 0 for FALSE and < 0 to omit the encoding so the
information is stored in its normal value. This does
however mean that we have to interpret the < 0 value
as FALSE in the relevant X509 code.