Implementation of X509_NAME using the new code.
This is probably the most complex type because we have
SEQUENCE OF { SET OF X509_NAME_ENTRY }
which would normally result in two nested STACKs.

However instead the data is stored in an X509_NAME structure which has a
single STACK_OF(X509_NAME_ENTRY) and a field in X509_NAME_ENTRY called
'set' which indicates the index of the SET OF structure for each entry.

It also caches the encoding for efficiency reasons.

This is implemented as an "external type" where the encoding and decoding
is done by explicit functions rather than via a template. We actually do
use an internal template though and convert the nested STACKs to X509_NAME
and back again.

The strange thing is this actually seems to work...