First major milestone of new code: replacement of all X509 certificate ASN1 routines.
Quite a bit of extra development on the base code was needed
to do this. In particular a new ASN1_AUX structure has been
added to carry miscellaneous information which most types wont
need.

Currently it includes a set of flags and information about the
reference count pointer.

It also has a callback function which is called at various
points in the ASN1 processing. This allows minor tweaks to be
performed such as caching data, freeing up extra fields and
various miscellaneous operations which can't be handled by
the core code. It is intended for minor customisation: for
major variations the use of external types is recommended.

And it still seems to work...